model segmentation 전체듣기
