+++ src/main/Test.kt
@@ -1,9 +1,6 @@
 package main
 
-import model.BugsAlbumCommentParser
-import model.ChartData
-import model.GenieAlbumCommentParser
-import model.MelonAlbumCommentParser
+import model.commentParser.MelonAlbumCommentParser
 import org.openqa.selenium.chrome.ChromeDriver
 
 fun main() {
