+++ src/model/commentParser/CrwalCommentTest.kt
@@ -1,20 +1,23 @@
-package model
+package model.commentParser
 
+import model.ChartData
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
 import org.openqa.selenium.WebDriver
 import org.openqa.selenium.chrome.ChromeDriver
 
 fun main() {
-	val t = TestParser()
-	t.crawl()
+	ChartData.getS_instance().bugsChartParser.chartDataParsing(null);
+	val c = CrwalCommentTest()
+	c.crawl()
 }
 
-class TestParser {
+class CrwalCommentTest {
+	//Properties
 	private var driver: WebDriver? = null
 	private val WEB_DRIVER_ID = "webdriver.chrome.driver"
 	private val WEB_DRIVER_PATH = "src/driver/chromedriver.exe"
-	private var base_url: String? = "https://www.melon.com/album/detail.htm?albumId=10479150"
+	private var base_url: String? = "https://music.bugs.co.kr/album/20356321?wl_ref=list_tr_11_chart"
 
 	init {
 		System.setProperty(WEB_DRIVER_ID, WEB_DRIVER_PATH)
@@ -28,16 +31,19 @@ class TestParser {
 		}
 		try {
 			//get page (= 브라우저에서 url을 주소창에 넣은 후 request 한 것과 같다)
-			base_url = 	"https://www.genie.co.kr/detail/albumInfo?axnm=81545399"
 			driver!!.get(base_url)
-			var html = driver!!.pageSource
-			var doc: Document = Jsoup.parseBodyFragment(html)
-			var arr = doc.select("div.reply-text > p")
-			println(arr)
+			val html = driver!!.pageSource
+			val doc: Document = Jsoup.parseBodyFragment(html)
+			val arr = doc.select("p[name=\"comment\"]").toMutableList()
+
+			for(ele in arr)
+				println(ele)
+
 		} catch (e: Exception) {
 			e.printStackTrace()
 		} finally {
 			driver!!.close()
 		}
 	}
+
 }
