+++ src/model/commentParser/TestParser.kt
@@ -1,4 +1,4 @@
-package model
+package model.commentParser
 
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
@@ -6,18 +6,15 @@ import org.openqa.selenium.WebDriver
 import org.openqa.selenium.chrome.ChromeDriver
 
 fun main() {
-	ChartData.getS_instance().bugsChartParser.chartDataParsing(null);
-	val c = CrwalCommentTest()
-	c.crawl()
+	val t = TestParser()
+	t.crawl()
 }
 
-class CrwalCommentTest {
-	//Properties
+class TestParser {
 	private var driver: WebDriver? = null
 	private val WEB_DRIVER_ID = "webdriver.chrome.driver"
 	private val WEB_DRIVER_PATH = "src/driver/chromedriver.exe"
-	val albumnum = "81743752"
-	private var base_url: String? = "https://music.bugs.co.kr/album/20356321?wl_ref=list_tr_11_chart"
+	private var base_url: String? = "https://www.melon.com/album/detail.htm?albumId=10479150"
 
 	init {
 		System.setProperty(WEB_DRIVER_ID, WEB_DRIVER_PATH)
@@ -31,19 +28,16 @@ class CrwalCommentTest {
 		}
 		try {
 			//get page (= 브라우저에서 url을 주소창에 넣은 후 request 한 것과 같다)
+			base_url = 	"https://www.genie.co.kr/detail/albumInfo?axnm=81545399"
 			driver!!.get(base_url)
-			val html = driver!!.pageSource
-			val doc: Document = Jsoup.parseBodyFragment(html)
-			val arr = doc.select("p[name=\"comment\"]").toMutableList()
-
-			for(ele in arr)
-				println(ele)
-
+			var html = driver!!.pageSource
+			var doc: Document = Jsoup.parseBodyFragment(html)
+			var arr = doc.select("div.reply-text > p")
+			println(arr)
 		} catch (e: Exception) {
 			e.printStackTrace()
 		} finally {
 			driver!!.close()
 		}
 	}
-
 }
